mcp_server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
pydantic
pyjwt
requests
